More plotting functions migrated
